Task 2: refactor routing
